LayoutTests/fast/dom/window-inner-size-scaling.html

<html>
    <script src="../../resources/js-test-pre.js"></script>
    <script>
    description("This test ensures window.innerWidth/innerHeight return the size of the visual viewport in CSS pixels.");
    var originalWidth = window.innerWidth;
    var originalHeight = window.innerHeight;
    var scale = 2;
    if (window.testRunner)
        window.testRunner.waitUntilDone();
    async function runTest() {
        if (window.testRunner)
            await window.testRunner.setPageScaleFactor(scale, 0, 0);

        shouldBe("window.innerWidth", "Math.floor(originalWidth / scale)");
        shouldBe("window.innerHeight", "Math.floor(originalHeight / scale)");
        if (window.testRunner)
            window.testRunner.notifyDone();
    }
    </script>
    <script src="../../resources/js-test-post.js"></script>
    <body onload="runTest()"></body>
</html>
