Safari-611.1.10/LayoutTests/fast/dom/window-inner-size-scaling-expected.txt

This test ensures window.innerWidth/innerHeight return the size of the visual viewport in CSS pixels.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.innerWidth is Math.floor(originalWidth / scale)
PASS window.innerHeight is Math.floor(originalHeight / scale)
PASS successfullyParsed is true

TEST COMPLETE
